fix navbar footer for mobile & pc view
- also fixed background image
