+++ client/src/components/chat/ChatSidebar.tsx
@@ -4,7 +4,7 @@ import SearchInput from '../reusables/SearchInput';
 
 function ChatSidebar() {
     return (
-        <div className="flex flex-col min-w-100 relative">
+        <div className="flex flex-col lg:ml-[120px] min-w-120 relative">
             <SearchInput onChange={(e) => {}} name="search" type="text" placeholder="Search users, groups" />
             <RoomList />
         </div>
