+++ client/src/styles/globals.css
@@ -36,7 +36,7 @@ body {
     font-family: 'Poppins';
     min-height: 100svh;
     background: url('../assets/images/background_desktop.png') no-repeat;
-    background-size: 100vw 100vh;
+    background-size: cover;
     background-position: center;
 }
 .chatsection {
@@ -53,6 +53,6 @@ body {
 @media screen and (max-width: 768px) {
     body {
         background: url('../assets/images/background_mobile.png') no-repeat;
-        background-size: 100vw 105vh;
+        background-size: cover
     }
 }
